Convert Circuit API to ZkFunction with ZkProgram-style structure
